refactor(persis): clean up queue store boundaries
