+++ implementation-notes.md
@@ -15,3 +15,6 @@
 - Wiring decision: changed `internal/cmd/context.go` to construct `store.NewQueueStore(file.NewCollection(cfg.Paths.QueueDir))`. I did not switch the broader context to a single shared `file.New(cfg.Paths.DataDir)` backend yet, because `dagrun`, `proc`, and distributed stores are still on legacy packages.
 - Test fixture wiring change: updated shared command/scheduler/API fixtures to construct the new queue store too. Keeping fixtures on `filequeue.New` created mixed-backend tests where the command deleted records through the new store but assertions read stale legacy queue indexes.
 - Scheduler package test tradeoff: package-local tests under `internal/service/scheduler` still instantiate `filequeue.New`. Importing `internal/persis/store` from that package creates an existing cycle because `store.WatermarkStore` imports scheduler types. Moving watermark or queue store packages apart would be a separate boundary cleanup.
+- Cleanup: `DequeueByName` and queue listing now use the `{queueName}/item_` prefix so legacy `.queue-index.json` files are never claimed as queue items.
+- Cleanup: moved `WatermarkStore` to `internal/persis/schedulerstore` so the generic `internal/persis/store` package no longer imports scheduler. Scheduler package tests can now use `store.NewQueueStore`.
+- Cleanup: split queue item parsing, cursor handling, and watcher polling into separate files. The legacy `internal/persis/filequeue` package is now explicitly documented as legacy-only coverage.
