+++ internal/intg/queue/fixture_test.go
@@ -21,7 +21,7 @@ import (
 	"github.com/dagucloud/dagu/internal/core/spec"
 	"github.com/dagucloud/dagu/internal/persis/file"
 	"github.com/dagucloud/dagu/internal/persis/filedagrun"
-	persiststore "github.com/dagucloud/dagu/internal/persis/store"
+	"github.com/dagucloud/dagu/internal/persis/schedulerstore"
 	"github.com/dagucloud/dagu/internal/runtime/transform"
 	"github.com/dagucloud/dagu/internal/service/scheduler"
 	"github.com/dagucloud/dagu/internal/test"
@@ -525,7 +525,7 @@ func (f *fixture) seedWatermark(lastTick, lastScheduledTime time.Time) {
 
 	wmBackend, err := file.New(f.th.Config.Paths.DataDir)
 	require.NoError(f.t, err)
-	store := persiststore.NewWatermarkStore(wmBackend.Collection("scheduler"))
+	store := schedulerstore.NewWatermarkStore(wmBackend.Collection("scheduler"))
 	state := &scheduler.SchedulerState{
 		Version:  1,
 		LastTick: lastTick,
